+++ resources/views/layouts/app.blade.php
@@ -17,6 +17,7 @@
 </head>
 <body class="min-h-screen bg-slate-50 text-slate-900 antialiased">
     <x-impersonation-banner />
+    <x-trial-banner />
     @include('layouts.navigation')
 
     @isset($header)